refactor: 디렉토리 계층 추가 및 권한 설정
- domain과 global 상위 디렉토리 추가하여 코드 구조화
- 모든 권한은 임시로 ROLE_USER으로 설정
